requirements.txt
streamlit
st-pages
pandas_profiling
streamlit-pandas-profiling